+++ cooked_input/convertors.py
@@ -9,6 +9,8 @@ import dateparser
 import csv
 from io import StringIO
 from future.utils import raise_from
+from abc import ABCMeta, abstractmethod
+
 
 from .error_callbacks import ConvertorError
 
@@ -21,13 +23,14 @@ TABLE_ID_OR_VALUE = -1
 ###
 ### Convertors:
 ###
-# class Convertor(metaclass=ABCMeta):
 class Convertor(object):
     # Abstract base class for conversion classes
-    def __init__(self, value_error_str, **kwargs):
+    __metaclass__ = ABCMeta
+
+    def __init__(self, value_error_str):
         self.value_error_str = value_error_str
 
-    # @abstractmethod
+    @abstractmethod
     def __call__(self, value, error_callback, convertor_fmt_str):
         pass
 
@@ -39,12 +42,10 @@ class IntConvertor(Convertor):
     :param base:  the radix base to use for the int conversion (default=10). Legal values are 0 and 2-36. See
         the Python int function for more information.
     :param value_error_str: the error string to use when an improper value is input.
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, base=10, value_error_str='an integer number', **kwargs):
-
+    def __init__(self, base=10, value_error_str='an integer number'):
         self._base = base
-        super(IntConvertor, self).__init__(value_error_str, **kwargs)
+        super(IntConvertor, self).__init__(value_error_str)
 
     def __call__(self, value, error_callback, convertor_fmt_str):
         result = None
@@ -65,10 +66,9 @@ class FloatConvertor(Convertor):
     convert to a floating point number
 
     :param value_error_str: the error string to use when an improper value is input.
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, value_error_str='a float number', **kwargs):
-        super(FloatConvertor, self).__init__(value_error_str, **kwargs)
+    def __init__(self, value_error_str='a float number'):
+        super(FloatConvertor, self).__init__(value_error_str)
 
     def __call__(self, value, error_callback, convertor_fmt_str):
         result = None
@@ -90,13 +90,12 @@ class BooleanConvertor(Convertor):
     convert to a boolean value (True or False).
 
     :param value_error_str: the error string to use when an improper value is input.
-    :param kwargs: no kwargs are currently supported.
 
     **BooleanConvertor** returns `True` for input values: 't', 'true', 'y', 'yes', and '1'. **BooleanConvertor** returns
     `False` for input values: 'f', 'false', 'n', 'no', '0'.
     """
-    def __init__(self, value_error_str='true or false', **kwargs):
-        super(BooleanConvertor, self).__init__(value_error_str, **kwargs)
+    def __init__(self, value_error_str='true or false'):
+        super(BooleanConvertor, self).__init__(value_error_str)
 
     def __call__(self, value, error_callback, convertor_fmt_str):
         true_set = {'t', 'true', 'y', 'yes', '1'}
@@ -123,7 +122,6 @@ class ListConvertor(Convertor):
         (ala CSV library.)
     :param elem_convertor: the convertor function to use for each element of the list (e.g. IntConvertor converts each
         element of the list to an integer
-    :param kwargs: no kwargs are currently supported
 
     For example, the accept a list of integers separated by colons (':') and return it as a Python list of ints::
 
@@ -131,31 +129,32 @@ class ListConvertor(Convertor):
       lc = ListConvertor(delimiter=':', elem_convertor=IntConvertor())
       result = get_input(prompt=prompt_str, convertor=lc)
     """
-    def __init__(self, value_error_str='list of values', delimiter=',', elem_convertor=None, **kwargs):
-        self.delimeter = delimiter
-        self.elem_convertor = elem_convertor
-        super(ListConvertor, self).__init__(value_error_str, **kwargs)
+    def __init__(self, value_error_str='list of values', delimiter=',', elem_convertor=None):
+        self._delimeter = delimiter
+        self._elem_convertor = elem_convertor
+        super(ListConvertor, self).__init__(value_error_str)
+
 
     def __call__(self, value, error_callback, convertor_fmt_str):
         buffer = StringIO(value)
 
-        if self.delimeter is None:
+        if self._delimeter is None:
             dialect = csv.Sniffer().sniff(value)
             dialect.skipinitialspace = True
         else:
-            csv.register_dialect('my_dialect', delimiter=self.delimeter, quoting=csv.QUOTE_MINIMAL, skipinitialspace=True)
+            csv.register_dialect('my_dialect', delimiter=self._delimeter, quoting=csv.QUOTE_MINIMAL, skipinitialspace=True)
             dialect = csv.get_dialect('my_dialect')
 
         reader = csv.reader(buffer, dialect)
         lst = next(reader)
 
         try:
-            if self.elem_convertor:
-                converted_list = [self.elem_convertor(item, error_callback, convertor_fmt_str) for item in lst]
+            if self._elem_convertor:
+                converted_list = [self._elem_convertor(item, error_callback, convertor_fmt_str) for item in lst]
             else:
                 converted_list = lst
         except ConvertorError:
-            raise ConvertorError(self.elem_convertor.value_error_str)
+            raise ConvertorError(self._elem_convertor.value_error_str)
 
         return converted_list
 
@@ -168,14 +167,13 @@ class DateConvertor(Convertor):
     convert to a datetime.
 
     :param value_error_str: the error string to use when an improper value is input.
-    :param kwargs: no kwargs are currently supported.
 
     Converts the cleaned input to an datetime value. Dateparser is used for the parsing, allowing
     a lot of flexibility in how date input is entered (e.g. '12/12/12', 'October 1, 2015', 'today', or 'next Tuesday').
     For more information about dateparser see: `<https://dateparser.readthedocs.io/en/latest/>`_
     """
-    def __init__(self, value_error_str='a date', **kwargs):
-        super(DateConvertor, self).__init__(value_error_str, **kwargs)
+    def __init__(self, value_error_str='a date'):
+        super(DateConvertor, self).__init__(value_error_str)
 
     def __call__(self, value, error_callback, convertor_fmt_str):
         result = dateparser.parse(value)
@@ -194,14 +192,13 @@ class YesNoConvertor(Convertor):
     convert to 'yes' or 'no'.
 
     :param value_error_str: the error string to use when an improper value is input.
-    :param kwargs: no kwargs are currently supported.
 
     **YesNoConvertor** returns `yes` for input values: 'y', 'yes', 'yeah', 'yup', 'aye', 'qui', 'si', 'ja', 'ken',
     'hai', 'gee', 'da', 'tak', 'affirmative'. **YesNoConvertor** returns `no` for input values: 'n', 'no', 'nope',
     'na', 'nae', 'non', 'negatory', 'nein', 'nie', 'nyet', 'lo'.
     """
-    def __init__(self, value_error_str='yes or no', **kwargs):
-        super(YesNoConvertor, self).__init__(value_error_str, **kwargs)
+    def __init__(self, value_error_str='yes or no'):
+        super(YesNoConvertor, self).__init__(value_error_str)
 
     def __call__(self, value, error_callback, convertor_fmt_str):
         yes_set = {'y', 'yes', 'yeah', 'yup', 'aye', 'qui', 'si', 'ja', 'ken', 'hai', 'gee', 'da', 'tak', 'affirmative'}
@@ -225,7 +222,6 @@ class ChoiceConvertor(Convertor):
 
     :param value_dict:  a dictionary containing keys to map from and values to map to
     :param value_error_str: the error string to use when an improper value is input.
-    :param kwargs: no kwargs are currently supported.
 
     convert a value to it's return value in a dictionary. Can be used to map the row index from a table of values or to
     map multiple tags to a single choice.
@@ -236,9 +232,9 @@ class ChoiceConvertor(Convertor):
       choice_convertor = ci.ChoiceConvertor(value_dict=value_map)
       result = ci.get_input(convertor=choice_convertor, prompt='Pick a color (1 - red, 2 - green, 3 - blue)')
     """
-    def __init__(self, value_dict=(), value_error_str='a valid row number', **kwargs):
+    def __init__(self, value_dict=(), value_error_str='a valid row number'):
         self._choices = value_dict
-        super(ChoiceConvertor, self).__init__(value_error_str, **kwargs)
+        super(ChoiceConvertor, self).__init__(value_error_str)
 
     def __call__(self, value, error_callback, convertor_fmt_str):
         result = None