+++ cooked_input/validators.py
@@ -16,6 +16,8 @@ import string
 import re
 import collections
 import logging
+from abc import ABCMeta, abstractmethod
+
 
 from .error_callbacks import print_error, silent_error, DEFAULT_VALIDATOR_ERROR
 from .input_utils import put_in_a_list, isstring
@@ -136,13 +138,15 @@ def validate(value, validators, error_callback=print_error, validator_fmt_str=DE
 ####
 #### Validators:
 ####
-# class Validator(metaclass=ABCMeta): # introduced in Python 3
 class Validator(object):
     # Abstract base class for validation classes
-    def __init__(self, **kwargs):
+    __metaclass__ = ABCMeta
+
+    @abstractmethod
+    def __init__(self):
         pass
 
-    # @abstractmethod   # introduced in Python 3
+    @abstractmethod
     def __call__(self, value, error_callback, validator_fmt_str):
         pass
 
@@ -154,12 +158,10 @@ class LengthValidator(Validator):
 
     :param min_len: the minimum required length for the input. If None (the default), no minimum length is checked.
     :param max_len: the maximum required length for the input. If None (the default), no maximum length is checked.
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, min_len=None, max_len=None, **kwargs):
+    def __init__(self, min_len=None, max_len=None):
         self._min_len = min_len
         self._max_len = max_len
-        super(LengthValidator, self).__init__(**kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         try:
@@ -189,12 +191,9 @@ class EqualToValidator(Validator):
     check if a value is an exact value
 
     :param value: the value to match
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, value, **kwargs):
-
+    def __init__(self, value):
         self._value = value
-        super(EqualToValidator, self).__init__(**kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         condition1 = (self._value is None or value == self._value)
@@ -216,12 +215,10 @@ class RangeValidator(Validator):
 
     :param min_val: The minimum allowed value (i.e. value must be <= min_val). If None (the default), no minimum value is checked.
     :param max_val: The maximum allowed value (i.e. value must be >= min_val). If None (the default), no maximum value is checked.
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, min_val=None, max_val=None, **kwargs):
+    def __init__(self, min_val=None, max_val=None):
         self._min_val = min_val
         self._max_val = max_val
-        super(RangeValidator, self).__init__(**kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         try:
@@ -254,12 +251,10 @@ class ChoiceValidator(Validator):
     check if a value is in a list of choices. Note: if choices is mutable, it can be changed after the instance is created.
 
     :param choices: an iterable (tuple, list, or set) containing the allowed set of choice for the value.
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, choices, **kwargs):
+    def __init__(self, choices):
         # note: if choices is mutable, the choices can change after instantiation
         self._choices = put_in_a_list(choices)
-        super(ChoiceValidator, self).__init__(**kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         result = value in self._choices
@@ -280,11 +275,9 @@ class NoneOfValidator(Validator):
     check if a value is not in a set of validators. Note: if choices is mutable, it can be changed after the instance is created.
 
     :param validators: an iterable list of validators that should not match the input value.
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, validators, **kwargs):
+    def __init__(self, validators):
         self._validators = validators
-        super(NoneOfValidator, self).__init__(**kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         result = not_in(value, self._validators, error_callback, validator_fmt_str)
@@ -301,12 +294,10 @@ class AnyOfValidator(Validator):
     check if a value matches any of a set of validators (OR operation). Note: if choices is mutable, it can be changed after the instance is created.
 
     :param validators: an iterable list of validators. When the validators is called, return True once any of the validators matches.
-    :param kwargs: kwargs: no kwargs are currently supported.
 
     """
-    def __init__(self, validators, **kwargs):
+    def __init__(self, validators):
         self._validators = validators
-        super(AnyOfValidator, self).__init__(**kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         result = in_any(value, self._validators, error_callback, validator_fmt_str)
@@ -320,11 +311,10 @@ class IsFileValidator(Validator):
     """
     check is a string is the name of an existing filename
 
-    :param value - the filename to verify
-    :param kwargs: kwargs: no kwargs are currently supported.
+    :param value: the filename to verify
     """
-    def __init__(self, **kwargs):
-        super(IsFileValidator, self).__init__(**kwargs)
+    def __init__(self):
+        pass
 
     def __call__(self, value, error_callback, validator_fmt_str):
         if os.path.isfile(value):
@@ -343,28 +333,12 @@ class SimpleValidator(Validator):
     functions (e.g. validus validation functions see: [https://shopnilsazal.github.io/validus/].) Can also be used with func.partial [https://docs.python.org/3/library/functools.html]
     to wrap validation functions that take more complex parameters.
 
-    :param validators an iterable list of validators. When the validators is called, return True once any of the validators matches.
-    :param kwargs: kwargs: no kwargs are currently supported.
-
-    optional kwargs:
-
-        name: a string to use for the validator name in error messages
-
+    :param validators: an iterable list of validators. When the validators is called, return True once any of the validators matches.
+    :param name: an optional string to use for the validator name in error messages
     """
-    def __init__(self, validator_func, **kwargs):
+    def __init__(self, validator_func, name='SimpleValidator value'):
         self._validator = validator_func
         self._name = None
-
-        for k, v in kwargs.items():
-            if k == 'name':
-                self._name = '%s' % v
-            else:
-                logging.warning('Warning: SimpleValidator received unknown option (%s)' % k)
-
-        super_options_to_skip = {'name'}
-        super_kwargs = {k: v for k, v in kwargs.items() if k not in super_options_to_skip}
-
-        super(SimpleValidator, self).__init__(**super_kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         result = self._validator(value)
@@ -382,27 +356,12 @@ class RegexValidator(Validator):
     """
     check if a value matches a regular expression.
 
-    :param pattern: the regular expression to match.
-    :param kwargs: kwargs: no kwargs are currently supported.
-    
-    options:
-    
-    regex_desc: a human readable string to use for the regex (used for error messages)
+    :param pattern: the regular expression to match
+    :param regex_desc: a human readable string to use for the regex (used for error messages)
     """
-    def __init__(self, pattern, **options):
+    def __init__(self, pattern, regex_desc=None):
         self._regex = pattern
-        self._regex_desc = pattern
-
-        for k, v in options.items():
-            if k == 'regex_desc':
-                self._regex_desc = v
-            else:
-                logging.warning('Warning: get_input received unknown option (%s)' % k)
-
-        super_options_to_skip = {'regex_desc'}
-        super_kwargs = {k: v for k, v in options.items() if k not in super_options_to_skip}
-
-        super(RegexValidator, self).__init__(**super_kwargs)
+        self._regex_desc = regex_desc
 
     def __call__(self, value, error_callback, validator_fmt_str):
         try:
@@ -437,71 +396,69 @@ class PasswordValidator(Validator):
     :param allowed: a string containing the allowed characters in the password. Defaults to upper and lower case ascii
         letters, plus digits, plus punctuation characters
     :param disallowed: a string containing characters not allowed in the password. Defaults to None
-    :param kwargs: kwargs: no kwargs are currently supported.
     """
     def __init__(self, min_len=None, max_len=None, min_lower=0, min_upper=0, min_digits=0, min_puncts=0,
-                 allowed=None, disallowed=None, **kwargs):
-        self.valid_chars = set(string.ascii_letters + string.digits + string.punctuation)
-        self.min_len = min_len
-        self.max_len = max_len
-        self.min_lower = min_lower
-        self.min_upper = min_upper
-        self.min_digits = min_digits
-        self.min_puncts = min_puncts
+                 allowed=None, disallowed=None):
+        self._valid_chars = set(string.ascii_letters + string.digits + string.punctuation)
+        self._min_len = min_len
+        self._max_len = max_len
+        self._min_lower = min_lower
+        self._min_upper = min_upper
+        self._min_digits = min_digits
+        self._min_puncts = min_puncts
 
         if disallowed is not None:
-            self.disallowed = set(disallowed)
+            self._disallowed = set(disallowed)
         else:
-            self.disallowed = set()
+            self._disallowed = set()
 
         if allowed is not None:
-            self.valid_chars = set(allowed)
+            self._valid_chars = set(allowed)
 
-        self.valid_chars -= self.disallowed
-        super(PasswordValidator, self).__init__(**kwargs)
+        self._valid_chars -= self._disallowed
 
     def __call__(self, value, error_callback, validator_fmt_str):
         if isstring(value) is False:
             print('PasswordValidator: value "{}" is not a string.'.format(value), file=sys.stderr)
             return False
 
-        if len(set(value) - self.valid_chars):
+        if len(set(value) - self._valid_chars):
             error_callback(validator_fmt_str, 'password', 'cannot contain any of the following characters: {}'.format(
-                               set(value) - self.valid_chars))
+                               set(value) - self._valid_chars))
             return False
 
-        if self.min_len is not None and (len(value)) < self.min_len:
+        if self._min_len is not None and (len(value)) < self._min_len:
             error_callback(validator_fmt_str, 'password',
-                           'too short (minimum length is {})'.format(self.min_len))
+                           'too short (minimum length is {})'.format(self._min_len))
             return False
 
-        if self.max_len and len(value) > self.max_len:
-            error_callback(validator_fmt_str, 'password', 'too long (maximum length is {})'.format(self.max_len))
+        if self._max_len and len(value) > self._max_len:
+            error_callback(validator_fmt_str, 'password', 'too long (maximum length is {})'.format(self._max_len))
             return False
 
-        if self.min_lower and len([c for c in value if c in string.ascii_lowercase]) < self.min_lower:
+        if self._min_lower and len([c for c in value if c in string.ascii_lowercase]) < self._min_lower:
             error_callback(validator_fmt_str, 'password',
-                           'too few lower case characters (minimum is {})'.format(self.min_lower))
+                           'too few lower case characters (minimum is {})'.format(self._min_lower))
             return False
 
-        if self.min_upper and len([c for c in value if c in string.ascii_uppercase]) < self.min_upper:
-            error_callback(validator_fmt_str, 'password', 'too few upper case characters (minimum is {})'.format(self.min_upper))
+        if self._min_upper and len([c for c in value if c in string.ascii_uppercase]) < self._min_upper:
+            error_callback(validator_fmt_str, 'password', 'too few upper case characters (minimum is {})'.format(self._min_upper))
             return False
 
-        if self.min_digits and len([c for c in value if c in string.digits]) < self.min_digits:
-            error_callback(validator_fmt_str, 'password', 'too few digit characters (minimum is {})'.format(self.min_digits))
+        if self._min_digits and len([c for c in value if c in string.digits]) < self._min_digits:
+            error_callback(validator_fmt_str, 'password', 'too few digit characters (minimum is {})'.format(self._min_digits))
             return False
 
-        if self.min_puncts and len([c for c in value if c in string.punctuation]) < self.min_puncts:
-            error_callback(validator_fmt_str, 'password', 'too few punctuation characters (minimum is {})'.format(self.min_puncts,
-                           set(string.punctuation) - self.disallowed))
+        if self._min_puncts and len([c for c in value if c in string.punctuation]) < self._min_puncts:
+            error_callback(validator_fmt_str, 'password', 'too few punctuation characters (minimum is {})'.format(self._min_puncts,
+                                                                                                                  set(string.punctuation) - self._disallowed))
             return False
 
         return True
 
     def __repr__(self):
         return 'PasswordValidator(allowed=%r, min_len=%r, max_len=%r, min_lowercase=%r, min_uppercase=%r, min_digits=%r, min_puncts=%r)' %\
-               (self.valid_chars, self.min_len, self.max_len, self.min_lower, self.min_upper, self.min_digits, self.min_puncts)
+               (self._valid_chars, self._min_len, self._max_len, self._min_lower, self._min_upper, self._min_digits, self._min_puncts)
 
 
 class ListValidator(Validator):
@@ -510,12 +467,10 @@ class ListValidator(Validator):
 
     :param len_validator: a validator to run on the list length.
     :param elem_validators: a single or list of validators to apply to the elements of the list.
-    :param kwargs: no kwargs are currently supported.
     """
-    def __init__(self, len_validator=None, elem_validators=None, **kwargs):
+    def __init__(self, len_validator=None, elem_validators=None):
         self._len_validator = len_validator
         self._elem_validators = elem_validators
-        super(ListValidator, self).__init__(**kwargs)
 
     def __call__(self, value, error_callback, validator_fmt_str):
         if self._len_validator: